refactor: adjust files and folders
