+++ src/data_structures/errors/index.go
@@ -1,6 +1,5 @@
-package data_structures
+package ds_errors
 
-// Error message constants
 const (
 	IndexOutOfBoundsError = "index out of bounds"
 	DequeEmptyError       = "deque is empty"
@@ -8,4 +7,5 @@ const (
 	StackEmptyError       = "stack is empty"
 	UnexpectedError       = "unexpected error: %v"
 	ExpectedError         = "expected error, got nil"
+	ValueNotFoundError    = "value not found in the tree"
 )
