+++ tests/data_structures/linear/stack/stack_test.go
@@ -1,13 +1,14 @@
 package data_structures_linear_test
 
 import (
-	data_structures "dsa/src/data_structures/linear"
+	ds_errors "dsa/src/data_structures/errors"
+	ds "dsa/src/data_structures/linear/stack"
 	"testing"
 )
 
 // TestNewStack tests the creation of a new stack and its initial size and empty state.
 func TestNewStack(t *testing.T) {
-	stack := data_structures.NewStack() // Create a new empty stack.
+	stack := ds.NewStack() // Create a new empty stack.
 
 	// Test the initial size of the stack.
 	t.Run("Initial Size", func(t *testing.T) {
@@ -26,7 +27,7 @@ func TestNewStack(t *testing.T) {
 
 // TestStackPush tests the Push method of the stack.
 func TestStackPush(t *testing.T) {
-	stack := data_structures.NewStack() // Create a new empty stack.
+	stack := ds.NewStack() // Create a new empty stack.
 
 	t.Run("Push elements", func(t *testing.T) {
 		// Push elements onto the stack.
@@ -53,7 +54,7 @@ func TestStackPush(t *testing.T) {
 
 // TestStackPop tests the Pop method of the stack.
 func TestStackPop(t *testing.T) {
-	stack := data_structures.NewStack() // Create a new empty stack.
+	stack := ds.NewStack() // Create a new empty stack.
 	stack.Push(1)
 	stack.Push(2)
 	stack.Push(3)
@@ -62,7 +63,7 @@ func TestStackPop(t *testing.T) {
 		// Pop elements from the stack and check the values and size.
 		value, error := stack.Pop()
 		if error != nil {
-			t.Errorf(data_structures.UnexpectedError, error)
+			t.Errorf(ds_errors.UnexpectedError, error)
 		}
 		if value != 3 {
 			t.Errorf("expected value 3, got %v", value)
@@ -73,7 +74,7 @@ func TestStackPop(t *testing.T) {
 
 		value, error = stack.Pop()
 		if error != nil {
-			t.Errorf(data_structures.UnexpectedError, error)
+			t.Errorf(ds_errors.UnexpectedError, error)
 		}
 		if value != 2 {
 			t.Errorf("expected value 2, got %v", value)
@@ -84,7 +85,7 @@ func TestStackPop(t *testing.T) {
 
 		value, error = stack.Pop()
 		if error != nil {
-			t.Errorf(data_structures.UnexpectedError, error)
+			t.Errorf(ds_errors.UnexpectedError, error)
 		}
 		if value != 1 {
 			t.Errorf("expected value 1, got %v", value)
@@ -98,14 +99,14 @@ func TestStackPop(t *testing.T) {
 	t.Run("Pop from empty stack", func(t *testing.T) {
 		_, error := stack.Pop()
 		if error == nil {
-			t.Error(data_structures.ExpectedError)
+			t.Error(ds_errors.ExpectedError)
 		}
 	})
 }
 
 // TestStackPeek tests the Peek method of the stack.
 func TestStackPeek(t *testing.T) {
-	stack := data_structures.NewStack() // Create a new empty stack.
+	stack := ds.NewStack() // Create a new empty stack.
 	stack.Push(1)
 	stack.Push(2)
 	stack.Push(3)
@@ -114,7 +115,7 @@ func TestStackPeek(t *testing.T) {
 		// Peek at the top element of the stack without removing it.
 		value, error := stack.Peek()
 		if error != nil {
-			t.Errorf(data_structures.UnexpectedError, error)
+			t.Errorf(ds_errors.UnexpectedError, error)
 		}
 		if value != 3 {
 			t.Errorf("expected value 3, got %v", value)
@@ -131,7 +132,7 @@ func TestStackPeek(t *testing.T) {
 		stack.Pop()
 		_, error := stack.Peek()
 		if error == nil {
-			t.Error(data_structures.ExpectedError)
+			t.Error(ds_errors.ExpectedError)
 		}
 	})
 }
